Move test to NXevent_data test
