+++ tests/nxdetector_test.py
@@ -288,3 +288,47 @@ def test_event_data_field_dims_labels(nexus_group: Tuple[Callable, LoadFromNexus
     with resource(builder)() as f:
         detector = nexus.NXroot(f, loader)['entry/detector_0']
         assert detector['detector_number'].dims == ['detector_number']
+
+
+def test_nxevent_data_selection_yields_correct_pulses(
+        nexus_group: Tuple[Callable, LoadFromNexus]):
+    event_time_offsets = np.array([456, 743, 347, 345, 632, 23])
+    event_data = EventData(
+        event_id=np.array([1, 2, 3, 1, 3, 2]),
+        event_time_offset=event_time_offsets,
+        event_time_zero=np.array([
+            1600766730000000000, 1600766731000000000, 1600766732000000000,
+            1600766733000000000
+        ]),
+        event_index=np.array([0, 3, 3, 5]),
+    )
+
+    builder = NexusBuilder()
+    builder.add_event_data(event_data)
+    resource, loader = nexus_group
+
+    with resource(builder)() as f:
+        events = nexus.NXroot(f, loader)['entry/events_0']
+
+        class Load:
+            def __getitem__(self, select=...):
+                da = events[select]
+                return da.bins.size().values
+
+        assert np.array_equal(Load()[...], [3, 0, 2, 1])
+        assert np.array_equal(Load()['pulse', 0], 3)
+        assert np.array_equal(Load()['pulse', 1], 0)
+        assert np.array_equal(Load()['pulse', 3], 1)
+        assert np.array_equal(Load()['pulse', -1], 1)
+        assert np.array_equal(Load()['pulse', -2], 2)
+        assert np.array_equal(Load()['pulse', 0:0], [])
+        assert np.array_equal(Load()['pulse', 1:1], [])
+        assert np.array_equal(Load()['pulse', 1:-3], [])
+        assert np.array_equal(Load()['pulse', 3:3], [])
+        assert np.array_equal(Load()['pulse', -1:-1], [])
+        assert np.array_equal(Load()['pulse', 0:1], [3])
+        assert np.array_equal(Load()['pulse', 0:-3], [3])
+        assert np.array_equal(Load()['pulse', -1:], [1])
+        assert np.array_equal(Load()['pulse', -2:-1], [2])
+        assert np.array_equal(Load()['pulse', -2:], [2, 1])
+        assert np.array_equal(Load()['pulse', :-2], [3, 0])
